added mentors data into the data folder under mentor, removed some non-important code
